ui/requirements.txt
streamlit==1.41.1
requests==2.32.3
pillow==10.4.0
